Rename Scrivener specific logic.
Also enforces Scrivener.Page is passed in and only availab eif Scrivener
is a dep.
